Alert before deleting

+++ src/main/java/pl/edu/agh/diskstalker/controller/FolderDetailsController.java
@@ -3,10 +3,15 @@ package pl.edu.agh.diskstalker.controller;
 import com.google.inject.Inject;
 import javafx.event.ActionEvent;
 import javafx.fxml.FXML;
+import javafx.scene.control.Alert;
+import javafx.scene.control.ButtonType;
 import javafx.scene.control.TextField;
 import javafx.stage.Stage;
 import pl.edu.agh.diskstalker.database.model.Root;
 import pl.edu.agh.diskstalker.presenter.FolderDetailsHandler;
+
+import javax.swing.text.html.Option;
+import java.util.Optional;
 
 public class FolderDetailsController {
     private Stage dialogStage;
@@ -66,7 +71,14 @@ public class FolderDetailsController {
 
     @FXML
     private void  handleFolderDeleteAction(ActionEvent event) {
-        detailsHandler.deleteRoot(root);
+        var result  = showAlert("Folder will be deleted!", "Are you sure?");
+        if (result.get() == ButtonType.OK){
+            detailsHandler.deleteRoot(root);
+        } else {
+            // ... user chose CANCEL or closed the dialog
+            dialogStage.close();
+        }
+
         dialogStage.close();
     }
 
@@ -74,5 +86,15 @@ public class FolderDetailsController {
     private void handleFolderCleanAction(ActionEvent event) {
         detailsHandler.cleanRoot(root);
         dialogStage.close();
+    }
+
+    private Optional<ButtonType>  showAlert(String message,  String details){
+        Alert alert = new Alert(Alert.AlertType.CONFIRMATION);
+        alert.setTitle("Confirmation Dialog");
+        alert.setHeaderText(message);
+        alert.setContentText(details);
+
+        Optional<ButtonType> result = alert.showAndWait();
+        return result;
     }
 }
